+++ documentation/example/app/api_auth/app/src/options.py
@@ -1,5 +1,5 @@
 import package.lib as lib
 
 
-def main(event, context):
+def main(event: dict, context: dict) -> dict:
     return lib.basic_return_value(event)